refactor: unify the `run` functions of `all` and `any`
